auth-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.0
sqlalchemy==1.4.23
pymysql==1.0.2
cryptography==3.4.7 